requirements.txt
aiofiles==0.5.0
amqp==2.5.2
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
Beaker==1.11.0
billiard==3.6.3.0
blinker==1.4
braintree==4.1.0
cachetools==4.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.1
cryptography==2.9.2
cssselect2==0.3.0
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
Django==3.0.6
dnspython==1.16.0
easy-thumbnails==2.7
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
flower==0.9.4
future==0.18.2
google-api-core==1.17.0
google-auth==1.16.0
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-logging==1.15.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
html5lib==1.0.1
humanize==0.5.1
Hypercorn==0.9.5
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
lazy==1.4
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
oauthlib==3.1.0
packaging==20.4
pbr==5.4.5
Pillow==7.1.2
pipenv==2020.5.28
pluggy==0.13.1
ply==3.11
priority==1.3.0
protobuf==3.12.2
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
pyparsing==2.4.7
Pyphen==0.9.5
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
sqlparse==0.3.1
stevedore==1.32.0
supervisor==4.2.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.2
unicode==2.7
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
URLObject==2.4.3
vine==1.3.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.4
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==0.15.0
WTForms==2.3.1
youtube-dl==2020.5.29
zipp==3.1.0
